3 2 compatible

+++ utils/run.sh
@@ -1,16 +1,19 @@
 #!/usr/bin/env bash
 
-docker rm -f tron
+if [[ $(docker inspect tron | grep NetworkSettings) != "" ]]; then
+  docker rm -f tron
+fi
 
 docker run -it \
-  --rm \
   -p 8090:8090 \
   -p 8091:8091 \
   -p 8092:8092 \
   -p 9090:9090 \
+  -p 50051:50051 \
+  -p 50052:50052 \
   -e "defaultBalance=100000" \
   -e "showQueryString=true" \
   -e "showBody=true" \
   -e "formatJson=true" \
   --name tron \
-  tron-quickstart
+  tronquickstart
